+++ lolProject/lolLib/EF/participant.cs
@@ -22,7 +22,7 @@ namespace lolLib.EF
     
         public int participantId { get; set; }
         public long gameId { get; set; }
-        public Nullable<sbyte> teamId { get; set; }
+        public Nullable<short> teamId { get; set; }
         public Nullable<short> championId { get; set; }
         public Nullable<int> spell1Id { get; set; }
         public Nullable<int> spell2Id { get; set; }
